HDDS-4236. Move "Om*Codec.java" to new project hadoop-ozone/interface-storage (#1424)
