+++ hadoop-ozone/interface-storage/src/main/java/org/apache/hadoop/ozone/om/codec/package-info.java
@@ -21,7 +21,4 @@
 /**
  * Utility classes to encode/decode DTO objects to/from byte array.
  */
-/**
- * OM DB definitions.
- */
 package org.apache.hadoop.ozone.om.codec;
