+++ hadoop-ozone/interface-storage/src/main/java/org/apache/hadoop/ozone/om/package-info.java
@@ -19,9 +19,6 @@
  */
 
 /**
- * Utility classes to encode/decode DTO objects to/from byte array.
+ * OMMetadataManager.
  */
-/**
- * OM DB definitions.
- */
-package org.apache.hadoop.ozone.om.codec;
+package org.apache.hadoop.ozone.om;
